backend/requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0          # [standard] adds watchfiles + websockets
starlette==0.37.2                  # pinned to match fastapi

# ── Database ──────────────────────────────────────────────────────────────────
psycopg2-binary==2.9.9
sqlalchemy==2.0.30

# ── Configuration ─────────────────────────────────────────────────────────────
python-dotenv==1.0.1
pydantic==2.7.1

# ── AI ────────────────────────────────────────────────────────────────────────
anthropic==0.25.0

# ── Gmail (OAuth2 + API client) ───────────────────────────────────────────────
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.127.0

# ── WhatsApp ──────────────────────────────────────────────────────────────────
httpx==0.27.0

# ── Kafka (async) ─────────────────────────────────────────────────────────────
aiokafka==0.10.0

# ── Logging ───────────────────────────────────────────────────────────────────
python-json-logger==2.0.7

# ── Error tracking (optional — remove if you don't use Sentry) ───────────────
sentry-sdk[fastapi]==2.3.1

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0                      # needed by fastapi TestClient
